move existing reducer code into a new folder
